requirements.txt
bayesian_optimization==1.4.3
Flask==2.2.5
Flask_RESTful==0.3.10
keras==2.9.0
matplotlib==3.5.2
numpy==1.23.3
pandas==1.4.4
Requests==2.31.0
scikit_learn==1.2.2
tensorflow==2.9.1
psycopg2-binary==2.9.6
flask-sqlalchemy==3.0.5



#pandas==1.4.4
#matplotlib==3.7.1
#numpy==1.23.1
#tensorflow==2.7.0
#Keras==2.7.0
#scikit-learn==1.2.2
#bayesian-optimization==1.2.0
#Flask==2.2.4
#Flask-RESTful==0.3.10
#requests==2.26.0





#Flask==2.2.4
#flask_restful==0.3.10
#keras==2.7.0
#matplotlib==3.7.1
#numpy==1.23.1
#pandas==1.4.4
#Requests==2.31.0
#scikit_learn==1.3.4
#tensorflow==2.7.0
